+++ src/main/java/com/example/catalogue/domain/productService/model/product/Measure.java
@@ -1,8 +1,8 @@
 package com.example.catalogue.domain.productService.model.product;
 
 public record Measure (
-    int numberOfItemsInPack,
-    MeasureType measureType,
-    int measureCount,
-    String unitOfMeasure
+        int numberOfItemsInPack,
+        MeasureType measureType,
+        Long measureCount,
+        String unitOfMeasure
 ) {};
